problem arret client

+++ client.c
@@ -6,15 +6,12 @@
 /*   By: bducrocq <bducrocq@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/22 12:05:16 by bducrocq          #+#    #+#             */
-/*   Updated: 2022/03/14 17:13:49 by bducrocq         ###   ########.fr       */
+/*   Updated: 2022/03/14 19:01:30 by bducrocq         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "includes/minitalk.h"
 
-enum e_bool	msg_ok;
-
-//TODO:	-debugger sinal
 /**
  * @brief send message in binary
  * 
@@ -28,16 +25,16 @@ void	ft_sendbin(char *str, int pid, int speed)
 	char	*tmp;
 	int		j;
 	int		i;
-	
+
 	i = 0;
 	while (str[i])
 	{
 		tmp = ft_atob(str[i]);
 		j = 0;
-		while(tmp[j])
+		while (tmp[j])
 		{
 			if (tmp[j] == '0')
-				kill(pid, SIGUSR1); 
+				kill(pid, SIGUSR1);
 			if (tmp[j] == '1')
 				kill(pid, SIGUSR2);
 			usleep(speed);
@@ -59,28 +56,36 @@ void	ft_sendbin(char *str, int pid, int speed)
 char	*ft_imax_to_str(int size, int count)
 {
 	t_tools	to;
-	
+
 	to.str = ft_calloc(count + 1, sizeof(char));
 	if (!to.str)
-		return(NULL);
+		return (NULL);
 	to.y = count - 1;
 	ft_memset(to.str, '0', count);
 	to.tmp = ft_itoa(size);
 	to.i = ft_strlen(to.tmp) - 1;
-	 while (to.i >= 0)
+	while (to.i >= 0)
 		to.str[to.y--] = to.tmp[to.i--];
 	free(to.tmp);
-	return(to.str);
+	return (to.str);
 }
 
-void signal_handler_client(int signal)
+void	signal_handler_client(int signal)
 {
-	if (signal == SIGUSR1)
+	static int	i = 0;
+
+	if (signal == SIGUSR1 && i >= 10)
 	{
-		ft_putstr_fd("The server has received the message\n", 1);
-		msg_ok = TRUE;
+		g_msg_ok = TRUE;
+		ft_putstr_fd("The server has received the message2\n", 1);
 		kill(getpid(), SIGQUIT);
 	}
+}
+
+void	initialize_client(t_tools	*ptc)
+{
+	ptc->bool = 0;
+	g_msg_ok = FALSE;
 }
 
 int	main(int ac, char **av)
@@ -88,17 +93,12 @@ int	main(int ac, char **av)
 	t_tools	tc;
 	t_tools	*ptc;
 	int		pidserv;
-	
+
 	if (ac != 3)
-	{
-		printf("Please enter PID SERVER and \"MESSAGE\"\n");
-		return (0);
-	}
-	printf("pid client: %d\n", getpid());
-	ptc = &tc;
-	ptc->bool = 0;
-	msg_ok = FALSE;
+		return (printf("Please enter PID SERVER and \"MESSAGE\"\n"));
 	pidserv = ft_atoi(av[1]);
+	ptc = &tc;
+	initialize_client(ptc);
 	ptc->str = ft_imax_to_str(ft_strlen(av[2]), 10);
 	if (!(ptc->str))
 		return (0);
@@ -111,7 +111,8 @@ int	main(int ac, char **av)
 	free(ptc->str);
 	ft_sendbin(av[2], pidserv, SPEED);
 	signal(SIGUSR1, signal_handler_client);
-	// while(1)
-	// 	pause();
+	while (g_msg_ok != TRUE)
+		pause();
+	ft_putstr_fd("The server has received the message3\n", 1);
 	return (0);
 }
